requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
dateparser==1.2.1
Django==5.0.6
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.15.2
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.41
idna==3.10
multidict==6.1.0
postgres==4.0
propcache==0.2.1
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pycryptodome==3.21.0
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
setuptools==69.0.3
six==1.17.0
smmap==5.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzlocal==5.3
urllib3==2.3.0
websockets==15.0
whitenoise==6.7.0
yarl==1.18.3
